Task 3: Undo add and remove from reading list
Task
Update the code such that a snackbar appears whenever the user adds or removes a book. (Hint: this module is already installed and set up)
The snackbar must display the event that occurred (i.e. added/removed), and an Undo action.
When the user clicks Undo it should set the reading list state back to the previous state.